add profile page and reformat homepage date

+++ index.php
@@ -4,11 +4,16 @@ include 'connection.php';
 
 session_start();
 
-$query = 'SELECT id, title, content, date, color, created_at FROM post_it ORDER BY created_at DESC';
-$response = $bdd->query($query);
-$datas = $response->fetchAll();
+if (isset($_SESSION['user'])) {
+    $query = 'SELECT p.id, p.user_id, p.title, p.content, UNIX_TIMESTAMP(p.date) AS date, p.color, p.created_at FROM post_it AS p INNER JOIN users AS u ON u.id = p.user_id WHERE u.email=:email ORDER BY created_at DESC';
+    $response = $bdd->prepare($query);
+    $response->execute([
+        'email' => $_SESSION['user']['email']
+    ]);
+    $datas = $response->fetchAll();
 
-$response->closeCursor();
+    $response->closeCursor();
+}
 
 ?>
 
@@ -26,22 +31,26 @@ $response->closeCursor();
     <section class="container">
         <div class="postit-list">
 
-            <?php foreach ($datas as $data) { ?>
-                <article class="postit" style="background-color: <?= $data['color'] ?>;">
-                    <div class="postit-header">
-                        <h2>
-                            <?= $data['title'] ?>
-                        </h2>
-                        <a href="delete.php?id=<?= $data['id'] ?>" title="Supprimer le post-it"><img width="30" height="30"
-                                src="https://img.icons8.com/sf-black-filled/64/cancel.png" alt="remove" /></a>
-                    </div>
-                    <p>
-                        <?= $data['content'] ?>
-                    </p>
-                    <p>
-                        <?= $data['date'] ?>
-                    </p>
-                </article>
+            <?php if (isset($_SESSION['user'])) { ?>
+                <?php foreach ($datas as $data) { ?>
+                    <article class="postit" style="background-color: <?= $data['color'] ?>;">
+                        <div class="postit-header">
+                            <h2>
+                                <?= $data['title'] ?>
+                            </h2>
+                            <a href="delete.php?id=<?= $data['id'] ?>" title="Supprimer le post-it"><img width="30" height="30"
+                                    src="https://img.icons8.com/sf-black-filled/64/cancel.png" alt="remove" /></a>
+                        </div>
+                        <p>
+                            <?= $data['content'] ?>
+                        </p>
+                        <p>
+                            <?= date('d/m/Y', $data['date']) ?>
+                        </p>
+                    </article>
+                <?php } ?>
+            <?php } else { ?>
+                <h2>Connectez-vous pour voir vos post-it ou en ajouter.</h2>
             <?php } ?>
 
         </div>
